# Reduce memory usage in `Oj.load` with using Ruby 3.0 feature