--- a/ext/oj/strict.c
+++ b/ext/oj/strict.c
@@ -35,18 +35,22 @@ VALUE oj_calc_hash_key(ParseInfo pi, Val parent) {
         if (Yes == pi->options.sym_key) {
             rkey = ID2SYM(rb_intern3(parent->key, parent->klen, oj_utf8_encoding));
         } else {
+#if HAVE_RB_ENC_INTERNED_STR
+            rkey = rb_enc_interned_str(parent->key, parent->klen, oj_utf8_encoding);
+#else
             rkey = rb_str_new(parent->key, parent->klen);
             rkey = oj_encode(rkey);
+            OBJ_FREEZE(rkey);
+#endif
         }
-        OBJ_FREEZE(rkey);
         return rkey;
     }
     if (Yes == pi->options.sym_key) {
 	rkey = oj_sym_intern(parent->key, parent->klen);
     } else {
 	rkey = oj_str_intern(parent->key, parent->klen);
-    }
     OBJ_FREEZE(rkey);
+    }
     return rkey;
 }