`rb_enc_interned_str()` was added in Ruby 3.0 which returns frozen string object.

It has same effect with `# frozen_string_literal: true` magic comment and
The object will be reused and reduce memory usage if you create a frozen String
with the same C string over and over.

Here is memory usage with sample code.

−               | before   | after
--               | --       | --
Oj.dump          | 33.904k  | 25.904k

### Environment
- MacBook Air (M1, 2020)
- macOS 12.0 beta 4
- Apple M1
- Ruby 3.0.2

### Before
```
[CPU]====================
Warming up --------------------------------------
             Oj.load     3.855k i/100ms
Calculating -------------------------------------
             Oj.load     38.190k (± 0.7%) i/s -    192.750k in   5.047359s
[MEMORY]====================
Calculating -------------------------------------
             Oj.load    33.904k memsize (     0.000  retained)
                       502.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)
```

### After
```
[CPU]====================
Warming up --------------------------------------
             Oj.load     3.758k i/100ms
Calculating -------------------------------------
             Oj.load     37.368k (± 0.9%) i/s -    187.900k in   5.028814s
[MEMORY]====================
Calculating -------------------------------------
             Oj.load    25.904k memsize (     0.000  retained)
                       302.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)
```

### Test code
Before, following objects will be generated for one Hash object

- A Hash object
- Two String as Hash key
- Two String as Hash value

After, the same hash keys will be reused.
(The following objects will be generated for one Hash object with this patch)

- A Hash object
- Two String as Hash value

```ruby
require 'benchmark/ips'
require 'benchmark/memory'
require 'oj'

item = {
  "type" => "string",
  "description" => "The important things"
}

data = (1..100).map { item }
json = Oj.dump(data)

puts "[CPU]" + "=" * 20
Benchmark.ips do |x|
  x.report('Oj.load') { Oj.load(json) }
end

puts "[MEMORY]" + "=" * 20
Benchmark.memory do |x|
  x.report('Oj.load') { Oj.load(json) }
end
```