@ohler55 2021-08-21
Let's discuss this a bit. I did some benchmarks when I added the cache option and my cache ended up being faster so when someone select cached keys that cache is used. The reason it is optional is that when parsing something like a dictionary with 100,000 keys the caching locks up a large amount of memory that is is not collected during a GC. Calling the Ruby intern functions does the same thing I think. If it does then it is best not to intern or cache when the options is turned off. Can you verify that there is code in Ruby to expire the interned entries? I would expect not and if not then the Ruby intern functions should not be called when the caching option is off. Would you agree or have I missed something?