@ohler55 2021-08-21
You did give me an idea though. Oj make no claim that the keys or strings are interned forever but rather just as an optimization. Maybe I could set the cache to have a limited number of entries and then flush out less used entries to save space.