@Watson1978 2021-08-22
At least, looks for me that it checks wheather the frozen string has been garbaged when we will register the string in `fstr_update_callback()`
- https://github.com/ruby/ruby/blob/18031f4102d7c8efc3a40b3fa65b36f94fc8a418/string.c#L388-L408
- https://github.com/ruby/ruby/blob/18031f4102d7c8efc3a40b3fa65b36f94fc8a418/string.c#L299-L348

It is also used in several libraries and the cache will be used effectively across these libraries.

- json
  - https://github.com/flori/json/blob/5ed8cf7eb1e60eba81909336a540437fd74745d3/ext/json/ext/parser/parser.c#L2457
- msgpack-ruby
  - https://github.com/msgpack/msgpack-ruby/blob/77e75afe33c3908497c98e8916845527237bc29c/ext/msgpack/buffer.h#L459
- mysql2
  - https://github.com/brianmario/mysql2/blob/7f4e844fccf6afa888d0bd108d4707a2a7784484/ext/mysql2/result.c#L175