@ohler55 2021-08-22
I agree with your point that a shared intern cache is preferred. I'm not sure that is justification for either dropping the faster Oj cache or for interning all key strings. Even in the example you pointed out, the json gem only interns if the "intern" flag is true.

The `faster_update_callback` does show that frozen strings are collected at some point but not interned strings.

Stepping back lets look at what is desired. We want faster string creation or reuse on strings that are encountered frequently. Unless we can assure that the cache does lock up too much memory there should be some way of making sure that does not happen. Right now there is an option to turn on or off the caching of keys as well as shorter strings. That solution is not ideal but for now lets assume it should be use until we find a better way should as a auto-expiring cache. If the cache option is to be honoured then the Ruby intern functions should only be called when the cache option is true. That leaves us with which is better, the Oj cache or the Ruby intern. When I benchmarked the two options Oj gave better results. In favour of the Ruby intern though is that the intern cache is shared across all gems so depending on the gems in use the memory use may be lower. The choice then is which is preferred better performance or potential reduced memory use?

I have been giving some thought to ache control options other than a flag. Some possibilities are:
 * an auto-expiring cache based on most recently used
 * an option to flush the cache on GC
 * flush cache entries not used since the last GC
 * provide a call like `Oj.clear_cache`

Any of those options would allow the cache to always be on as the issue of unbounded memory use from huge dictionary files such as the one user tried that had over 100,000 keys in a Hash.