@Watson1978 2021-08-22
The objects cached by rb_enc_interned_str() are also garbaged with GC

Here is an easy experiment.

```diff
diff --git a/ext/oj/oj.c b/ext/oj/oj.c
index f01f656..d995db2 100644
--- a/ext/oj/oj.c
+++ b/ext/oj/oj.c
@@ -1748,6 +1748,13 @@ static VALUE protect_require(VALUE x) {
     return Qnil;
 }
 
+static VALUE oj_fstring(VALUE self, VALUE str) {
+#if HAVE_RB_ENC_INTERNED_STR
+    return rb_enc_interned_str(RSTRING_PTR(str), RSTRING_LEN(str), rb_utf8_encoding());
+#endif
+    rb_raise(rb_eRuntimeError, "Non supported Ruby version.");
+}
+
 /* Document-module: Oj
  *
  * Optimized JSON (Oj), as the name implies was written to provide speed
@@ -1809,6 +1816,8 @@ void Init_oj() {
     rb_define_module_function(Oj, "object_load", oj_object_parse, -1);
     rb_define_module_function(Oj, "wab_load", oj_wab_parse, -1);
 
+    rb_define_module_function(Oj, "fstring", oj_fstring, 1);
+
     rb_define_module_function(Oj, "dump", dump, -1);
 
     rb_define_module_function(Oj, "to_file", to_file, -1);
```

```ruby
require 'oj'

1_000_000.times do |i|
  Oj.fstring "key_#{i}"

  if i % 10000 == 0
    rss = Integer(`ps -o rss= -p #{Process.pid}`) / 1024.0
    puts "#{i},#{rss}"
  end
end
```

## Result
```
$ ruby memory.rb
0,14.73828125
10000,15.39453125
20000,15.62109375
30000,15.8515625
40000,15.890625
50000,15.90234375
60000,15.921875
70000,15.921875
80000,15.921875
90000,15.921875
100000,15.921875
110000,15.921875
120000,15.921875
130000,15.921875
140000,15.921875
150000,15.921875
160000,15.921875
170000,15.921875
180000,15.9296875
190000,15.9296875
200000,15.9296875
210000,15.9375
220000,15.9375
230000,15.9375
240000,15.9375
250000,15.9375
260000,15.9375
270000,15.9375
280000,15.9375
290000,15.9375
300000,15.9375
310000,15.9375
320000,15.9375
330000,15.9375
340000,15.9375
350000,15.9375
360000,15.9375
370000,15.9375
380000,15.9375
390000,15.9375
400000,15.9375
410000,15.9375
420000,15.9375
430000,15.9375
440000,15.9375
450000,15.9375
460000,15.9375
470000,15.9375
480000,15.9375
490000,15.9375
500000,15.9375
510000,15.9375
520000,15.9375
530000,15.9375
540000,15.9375
550000,15.9375
560000,15.9375
570000,15.9375
580000,15.9375
590000,15.9375
600000,15.9375
610000,15.9375
620000,15.9375
630000,15.9375
640000,15.9375
650000,15.9375
660000,15.9375
670000,15.9375
680000,15.9375
690000,15.9375
700000,15.9375
710000,15.9375
720000,15.9375
730000,15.9375
740000,15.9375
750000,15.9375
760000,15.9375
770000,15.9375
780000,15.9375
790000,15.9375
800000,15.9375
810000,15.9375
820000,15.94921875
830000,15.94921875
840000,15.94921875
850000,15.94921875
860000,15.94921875
870000,15.94921875
880000,15.94921875
890000,15.94921875
900000,15.94921875
910000,15.94921875
920000,15.94921875
930000,15.94921875
940000,15.94921875
950000,15.94921875
960000,15.94921875
970000,15.94921875
980000,15.94921875
990000,15.94921875
```

If the object created by rb_enc_interned_str() is persistent, it must be using more huge amount of memory.